+++ src/app/user-edit-add/user-edit-add.component.ts
@@ -17,6 +17,7 @@ import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 import { LoggerService } from '../services/logger/logger.service';
 import { PricingBaseComponent } from 'app/pricing/pricing-base/pricing-base.component';
+import { CachePuService } from 'app/services/cache/cache-pu.service';
 const swal = require('sweetalert');
 
 @Component({
@@ -151,6 +152,7 @@ export class UserEditAddComponent extends PricingBaseComponent implements OnInit
     public location: Location,
     public brandService: BrandService,
     private logger: LoggerService,
+    private cachePuService: CachePuService,
 
   ) {
     super(prjctPlanService, notify);
@@ -178,7 +180,6 @@ export class UserEditAddComponent extends PricingBaseComponent implements OnInit
     }
 
     this.getCurrentProject();
-    this.getAllUsersOfCurrentProject();
     this.getProjectPlan();
     this.getPendingInvitation();
     this.getBrowserLang();
@@ -397,19 +398,27 @@ export class UserEditAddComponent extends PricingBaseComponent implements OnInit
   }
 
   getAllUsersOfCurrentProject() {
+    // Clear the cache to ensure fresh data is always seen when entering the component
+    this.cachePuService.clearPuCache();
+    this.logger.log('[USER-EDIT-ADD] - GET ALL USERS OF CURRENT PROJECT - Cleared project users cache to ensure fresh data');
+    
     this.usersService.getProjectUsersByProjectId().subscribe((projectUsers: any) => {
-      this.logger.log('[USER-EDIT-ADD] - GET ALL PROJECT USERS OF THE PROJECT - RES ', projectUsers);
+     this.logger.log('[USER-EDIT-ADD] - GET ALL PROJECT USERS OF THE PROJECT - RES ', projectUsers);
 
-      if (projectUsers) {
+      if (projectUsers && projectUsers.length > 0) {
         this.projectUsersLength = projectUsers.length;
         this.logger.log('[USER-EDIT-ADD] - GET ALL PROJECT USERS OF THE PROJECT - PROJECT USERS Length ', this.projectUsersLength);
 
         const filteredProjectUser = projectUsers.filter((obj: any) => {
-          return obj.id_user._id === this.CURRENT_USER_ID;
+          return obj.id_user && obj.id_user._id === this.CURRENT_USER_ID;
         });
         this.logger.log('[USER-EDIT-ADD] - GET ALL PROJECT USERS OF THE PROJECT - filteredProjectUser FOR CURRENT_USER_ID ', filteredProjectUser);
-        this.currentUser_projectUserID = filteredProjectUser[0]._id
-
+        if (filteredProjectUser && filteredProjectUser.length > 0) {
+          this.currentUser_projectUserID = filteredProjectUser[0]._id;
+        }
+      } else {
+        this.logger.log('[USER-EDIT-ADD] - GET ALL PROJECT USERS OF THE PROJECT - Empty array or no project users');
+        this.projectUsersLength = 0;
       }
     }, error => {
       this.logger.error('[USER-EDIT-ADD] - GET ALL PROJECT USERS OF THE PROJECT - ERROR', error);
@@ -768,6 +777,8 @@ export class UserEditAddComponent extends PricingBaseComponent implements OnInit
         this.project_name = project.name;
         this.id_project = project._id;
         this.logger.log('[USER-EDIT-ADD] - GET CURRENT PROJECT - PROJECT-NAME ', this.project_name, ' PROJECT-ID ', this.id_project)
+        // Call getAllUsersOfCurrentProject() when project is available
+        this.getAllUsersOfCurrentProject();
       }
     });
   }